requirements.txt
Django==1.11.3
djangorestframework==3.6.3
requests==2.14.2
Scrapy==1.4.0
scrapy-djangoitem==1.1.1
six==1.10.0
psycopg2==2.7.1

celery==4.0.2
redis==2.10.5